feat: 장바구니에 상품 수량 기능 추가

+++ app/src/main/java/woowacourse/shopping/data/cart/LocalCartProductRepository.kt
@@ -45,9 +45,11 @@ class LocalCartProductRepository(
         totalCount++
     }
 
-    override fun delete(shoppingCartId: Long) {
+    override fun delete(shoppingCartId: Long?) {
         thread {
-            dao.delete(shoppingCartId)
+            if (shoppingCartId != null) {
+                dao.delete(shoppingCartId)
+            }
         }.join()
         totalCount--
     }
